+++ changes.md
@@ -6,9 +6,12 @@
 * Added soda can tower variant
   * Added CanTower (orchestrator class)
   * Added CanPlatform
+  * Added CanRail
   * Added can_tower example
   * Added can_platform example
+  * Added can_rail example
   * Added can_tower_set example
+ 
 
 ## 1.0.3
 * Upped cqterrain version to 1.0.3